WIP: keep the ABI-specific structs self-contained
Also, only move them in include/ for the ABI check, so that they don't
get shipped as public headers.
